Add takeback
Needed for makemove to work since parseMove only checks for psuedo legality and
the actually legality check (check, castle rights, etc.) are handled in makeMove.

There's currently a gnarly issue where e2e4 is not working as expected. I think
the path forward is to set up some unit tests, specifically for parseMove since
that is a core function that really has no room for error.