Fixup underAttack issue
Add tests
Update gitignore

---
The solution was genuinely just some print statements. The tests helped and so
did delve but really the prints saved me. Not sure why they worked in that
scenario. I happened to be running an interactive debugger on a test case but
not sure if that's what did it.